+++ changelog.md
@@ -5,6 +5,8 @@ The format of this file is based on [Keep a Changelog](https://keepachangelog.co
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] 
+### Changed
+- Modified release.sh so that it now updates the version in Dockerfile's LABEL step
 
 ## [1.0.1] - 2020-01-14
 ### Added